Move RunCommand to testing file
