+++ command/testing.go
@@ -1,12 +1,16 @@
 package command
 
 import (
+	"bytes"
 	"errors"
 	"fmt"
 	"reflect"
 
 	"github.com/AlecAivazis/survey/v2"
 	"github.com/AlecAivazis/survey/v2/core"
+	"github.com/google/shlex"
+	"github.com/spf13/cobra"
+	"github.com/spf13/pflag"
 
 	"github.com/cli/cli/api"
 	"github.com/cli/cli/context"
@@ -97,6 +101,52 @@ func initFakeHTTP() *api.FakeHTTP {
 		return api.NewClient(api.ReplaceTripper(http)), nil
 	}
 	return http
+}
+
+type cmdOut struct {
+	outBuf, errBuf *bytes.Buffer
+}
+
+func (c cmdOut) String() string {
+	return c.outBuf.String()
+}
+
+func (c cmdOut) Stderr() string {
+	return c.errBuf.String()
+}
+
+func RunCommand(cmd *cobra.Command, args string) (*cmdOut, error) {
+	rootCmd := cmd.Root()
+	argv, err := shlex.Split(args)
+	if err != nil {
+		return nil, err
+	}
+	rootCmd.SetArgs(argv)
+
+	outBuf := bytes.Buffer{}
+	cmd.SetOut(&outBuf)
+	errBuf := bytes.Buffer{}
+	cmd.SetErr(&errBuf)
+
+	// Reset flag values so they don't leak between tests
+	// FIXME: change how we initialize Cobra commands to render this hack unnecessary
+	cmd.Flags().VisitAll(func(f *pflag.Flag) {
+		switch v := f.Value.(type) {
+		case pflag.SliceValue:
+			_ = v.Replace([]string{})
+		default:
+			switch v.Type() {
+			case "bool", "string", "int":
+				_ = v.Set(f.DefValue)
+			}
+		}
+	})
+
+	_, err = rootCmd.ExecuteC()
+	cmd.SetOut(nil)
+	cmd.SetErr(nil)
+
+	return &cmdOut{&outBuf, &errBuf}, err
 }
 
 type errorStub struct {
